[DatasetTable|SorDropdown] sorting status preorder only on by_date sort | renamed tsx and css files, changed sort labels and fixed btn warning
