--- metaspace/webapp/src/components/SortDropdown/style.css
@@ -1,9 +1,0 @@
-.sort-dp-container {
-    width: 240px;
-}
-.sort-dp-btn {
-    margin-left: 2px;
-    border-left: 1px solid #BCCDDC;
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-}
